refactor: make code more std compliant

+++ src/strscpy.c
@@ -22,12 +22,12 @@
 
 #include "strscpy.h"
 
-ssize_t strscpy(char *dest, const char *src, size_t n)
+size_t strscpy(char *dest, const char *src, size_t n)
 {
     size_t i;
     for (i = 0; i < n; ++i){
         if ('\0' == (dest[i] = src[i])) 
-            return i > SSIZE_MAX ? -1 : (ssize_t) i;
+            return i > SSIZE_MAX ? -1 : i;
     }
 
     if (0 == i) return 0;
